fix avatar
